authorized requests done

+++ src/componets/dash/issueDetails.js
@@ -1,12 +1,20 @@
 import { useIssuesContext } from "../../hooks/useIssuesContext"
-
+import { useAuthContext } from "../../hooks/useAuthContext"
 const IssueDetails = ({issue}) => {
     const {dispatch} = useIssuesContext()
+    const {user} = useAuthContext()
     const { driver = [], route = [], room = [] } = issue
-
+    
     const handleClick = async () => {
+        if (!user){
+            return
+        }
+
         const response = await fetch('/api/issues/' + issue._id, {
-            method: "DELETE"
+            method: "DELETE",
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
